Change package name for sample 04.
