+++ AzureActiveDirectory/OAuth2/Servlet/sample-04-on-behalf-of-flow/sample-04-client/src/main/java/com/azure/spring/sample/active/directory/oauth2/servlet/sample04/client/ClientApplication.java
@@ -1,12 +1,12 @@
-package com.azure.spring.sample.active.directory.oauth2.sample04.resource.server1;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample04.client;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ResourceServerApplication {
+public class ClientApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(ResourceServerApplication.class, args);
+        SpringApplication.run(ClientApplication.class, args);
     }
 }
