+++ AzureActiveDirectory/OAuth2/Servlet/sample-04-on-behalf-of-flow/sample-04-client/src/main/java/com/azure/spring/sample/active/directory/oauth2/servlet/sample04/client/controller/HomeController.java
@@ -1,4 +1,4 @@
-package com.azure.spring.sample.active.directory.oauth2.sample04.resource.server1.controller;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample04.client.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -8,6 +8,6 @@ public class HomeController {
 
     @GetMapping("/")
     public String home() {
-        return "Hello, this is resource-server-1.";
+        return "Hello, this is sample-04-client.";
     }
 }
