+++ AzureActiveDirectory/OAuth2/Servlet/sample-04-on-behalf-of-flow/sample-04-client/src/main/java/com/azure/spring/sample/active/directory/oauth2/servlet/sample04/client/controller/ResourceServer1Controller.java
@@ -1,4 +1,4 @@
-package com.azure.spring.sample.active.directory.oauth2.sample04.client.controller;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample04.client.controller;
 
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClient;
 import org.springframework.security.oauth2.client.annotation.RegisteredOAuth2AuthorizedClient;
