+++ AzureActiveDirectory/OAuth2/Servlet/sample-04-on-behalf-of-flow/sample-04-resource-server-1/src/main/java/com/azure/spring/sample/active/directory/oauth2/servlet/sample04/resource/server1/configuration/ApplicationConfiguration.java
@@ -1,4 +1,4 @@
-package com.azure.spring.sample.active.directory.oauth2.sample04.resource.server1.configuration;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample04.resource.server1.configuration;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.security.oauth2.resource.OAuth2ResourceServerProperties;
