+++ AzureActiveDirectory/OAuth2/Servlet/sample-04-on-behalf-of-flow/sample-04-resource-server-2/src/main/java/com/azure/spring/sample/active/directory/oauth2/servlet/sample04/resource/server2/ResourceServerApplication.java
@@ -1,0 +1,12 @@
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample04.resource.server2;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ResourceServerApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ResourceServerApplication.class, args);
+    }
+}
